# back/swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "아임유어박스 RestAPI 문서정의,최초작성일:2020-11-02"
  version: "1.0.0"
  title: "아임유어박스 RESTAPI 정의서"
 
  contact:
    email: "markup3604@gmail.com"
  license:
    name: "ISC"
    url: "https://imyourbox.co.kr"
host: "127.0.0.1:4000"
basePath: "/order"
tags:
- name: "order"
  description: "아임유어박스 홈페이지 주문정보"
  externalDocs:
    description: "아임유어박스"
    url: "https://imyourbox.com"

schemes:
- "https"
- "http"
paths:
  /order_info:
    post:
      tags:
      - "order"
      summary: "이름,회사,이메일,핸드폰,요청사항만 받는 주문정보입니다"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "간단주문정보"
        required: true
        schema:
          $ref: "#/definitions/Order_Info"
      responses:
        "405":
          description: "Invalid input"
  /detail_order_info:
    post:
      tags:
      - "order"
      summary: "상세견적신청을 원하는 고객을 위한 주문정보"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Pet object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Detail_Order_Info"
      responses:
        "405":
          description: "Invalid input"
        
definitions:
  Order_Info:
    type: "object"
    properties:
      customer_company:
        type: "string"
        format: "100"
      customer_manager_name:
         type: "string"
         format: "100"
      customer_phone:
        type: "string"
        format: "100"
      customer_email:
        type: "string"
        format: "100"
      customer_memo:
        type: "string"
        format: "1000"
    xml:
      name: "Order"
  Detail_Order_Info:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"